avr: support `tinygo run/gdb` with simavr
This commit adds support for running both the Arduino Uno and the
DigiSpark under a simulator.

Running `tinygo gdb` under a simulator is different from (for example)
what `tinygo gdb` does for other boards such as the PCA10040. However, I
think this is the better option as probably very few people actually
have a hardware debugger and supporting GDB with the simulator is then
much more useful.